refactor: move ValidatedInput to the Input namespace
